# 🌍 #349 Improve integer generator run performance by 30-35%