We cache the work when we create an int gen, but inside a SelectMany, we're creating a new gen every iteration.

Lazily calculate the bounds ranges.

Before:

```
|                Method |      Mean |     Error |    StdDev |
|---------------------- |----------:|----------:|----------:|
|         SelectManyGen | 31.904 us | 0.5425 us | 0.8120 us |
| SelectManyGenBaseline |  9.950 us | 0.1692 us | 0.1500 us |
```

After:

```
|                Method |      Mean |     Error |    StdDev |
|---------------------- |----------:|----------:|----------:|
|         SelectManyGen | 20.077 us | 0.3833 us | 0.4101 us |
| SelectManyGenBaseline |  9.833 us | 0.1924 us | 0.2363 us |
```